refactor: move PCR pass to dedicated module
