+++ crates/revmc/src/bytecode/passes/mod.rs
@@ -5,6 +5,8 @@ pub(crate) use block_analysis::{Cfg, Snapshots};
 
 mod const_fold;
 
+mod pcr;
+
 mod dead_store_elim;
 
 mod dedup;
